requirements.txt
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
bcrypt==4.0.1
bleach==3.3.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==3.3.2
click==7.1.2
commonmark==0.9.1
cycler==0.10.0
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
et-xmlfile==1.1.0
extra-streamlit-components==0.1.56
gitdb==4.0.7
GitPython==3.1.18
idna==2.10
importlib-metadata==4.12.0
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.21.0
openpyxl==3.0.7
packaging==21.0
pandas==1.5.0
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pivottablejs==0.9.0
plotly==5.1.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.20.3
ptyprocess==0.7.0
pyarrow==17.0.0
pycparser==2.20
pydeck==0.9.1
Pygments==2.9.0
PyJWT==2.6.0
Pympler==1.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==6.0
pyzmq==22.1.0
requests==2.32.3
rich==12.5.1
semver==2.13.0
Send2Trash==1.7.1
simplejson==3.17.5
six==1.16.0
smmap==4.0.0
streamlit==1.38.0
streamlit-aggrid==0.2.1
streamlit-authenticator==0.2.1
tenacity==8.5.0
terminado==0.10.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
typing-extensions==4.3.0
tzlocal==2.1
urllib3==1.26.6
validators==0.18.2
watchdog==4.0.2
wcwidth==0.2.5
webencodings==0.5.1
wget==3.2
widgetsnbextension==3.5.1
zipp==3.8.1
